modified footstep audio
fixed some small bugs with the footstep audio volume increasing every time you jump.  Also changed jump logic so that y velocity is constant -0.1 when not jumping, instead of y remaining the same as the velocity at the end of your jump. Also made adjustments to stride intervals for footsteps to give a more weightless feel.