Created procedural nav mesh
added nav mesh baking in the code and altered wall generation so they are not all separate small sections to help avoid aliasing. made some slight modifications to the door look and added code to spawn enemies and the player in random room locations. added spotlights to spawn in each room and removed any other light sources from the scene