Added ability to put room prefabs into the procedurally generated level
Added the ability to create your own room prefab, attach a script and have it be reproduced as part of the procedurally generated level. There are still some bugs with corridors not connecting all rooms when more than 5 or so 20x20 rooms are added. Added some code to split the play area an amount approximate to the golden ration to give a less uniform grid pattern.
Also made modifications to the spotlight prefab to include a physical visible light, for which i also created a cookie for the spotlight so it appears as though it is projecting the correct shape to match the new light design. Fixed issues with light bleeding through floor and walls from other rooms.