Added Exit Room
Added exit room with plinths which require orbs of a specific colour to activate. After all 3 are activated the teleporter will start and be able to send you to the next level. On top of this I created "Orb Rooms" which contain plinths with the required colour of orb on them.